Add descriptor pool and sets
A descriptor pool is created and used to descriptor sets allocation.
There is a descriptor set for each MAX_FRAME_IN_FLIGHT.
The uniform buffers are associated with the descriptor sets and these are used in drawFrame() function.

Note: since the "proj" flips the Y axis (because a GLM problem using it for non-OpenGL), it is also necessary to change the frontFace definition in the graphics pipeline creation.