Fix minimization error
The use of SDL makes different the way to deal with minimization. The SDL_window returns its extent != 0 even if it is minimized, so you can't know if the window is shown up again through its extent. To do that, you have to wait for the SDL_EVENT_WINDOW_RESTORED, but this only has to be expected when window was minimized, it is not a condition for all cases of resizing.