Separate model and light UBOs
ModelUboManager manages all the uniform buffers of a single model. This is like this because the renderer will be able to render only a single model at a time.
LightUboManager manages all the uniform buffers of all the lights. A std::vector of uniform data about each light is stored for each frame.